+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Default `spec.location` field for CAPZ `AzureMachinePool` CRs.
+
 ## [0.3.0] - 2021-09-02
 
 ### Added